+++ src/lib/errors/DuplicatedRequestError.ts
@@ -1,0 +1,7 @@
+import {BaseError} from "./BaseError";
+
+export class DuplicatedRequestError extends BaseError {
+    constructor() {
+        super(`Request already exists.`);
+    }
+}