+++ src/lib/models/Request.ts
@@ -35,10 +35,23 @@ export class Request extends Model<Request> {
     @BelongsTo(() => User, "receiverId")
     public rReceiver: User;
 
+    /**
+     * Wrapper to return the request type for the request
+     */
+    public get type(): RequestType {
+        return this.requestType;
+    }
+
+    /**
+     * The receiver of the request
+     */
     public async receiver(): Promise<User> {
         return await this.$get("rReceiver") as User;
     }
 
+    /**
+     * The sender of the request.
+     */
     public async sender(): Promise<User> {
         return await this.$get("rSender") as User;
     }